requirements.txt
# Core application dependencies
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0

# AI and ML dependencies
openai>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Visualization dependencies
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation dependencies (optional)
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13